WIP: refactoring
big refactoring work: main goal is a more clean and stable code for
simulations of bending magnets, undulators and arbitrary magnetic
fields.

 Changes to be committed:
	new file:   src/barc4sr/aux_common.py
	modified:   src/barc4sr/aux_energy.py
	modified:   src/barc4sr/aux_plots.py
	modified:   src/barc4sr/aux_processing.py
	new file:   src/barc4sr/aux_rw.py
	modified:   src/barc4sr/aux_utils.py
        modified:   src/barc4sr/sr_bending_magnet.py
	modified:   src/barc4sr/sr_classes.py
