+++ src/barc4sr/aux_plots.py
@@ -151,43 +151,3 @@ def plot_magnetic_field(eBeamTraj: dict, direction: str, **kwargs) -> None:
     axes[-1].set_xlabel('[mm]', color='black')
     plt.tight_layout()
     plt.show()
-
-# def plot_magnetic_field(eBeamTraj_srw, **kwargs):
-#     xmin = kwargs.get("xmin", None)
-#     xmax = kwargs.get("xmax", None)
-#     ymin = kwargs.get("ymin", None)
-#     ymax = kwargs.get("ymax", None)
-
-#     integrals = kwargs.get("integrals", False)
-
-#     file_name = kwargs.get("file_name", None)
-
-#     img = PlotManager(eBeamTraj_srw["eTraj"]["Bx"],  eBeamTraj_srw["eTraj"]["Z"])
-#     img.additional_info('magnetic field', "s [m]",  "B [T]", xmin=xmin, xmax=xmax, ymin=ymin, ymax=ymax)
-#     img.aesthetics(400, True, True, 0, 1, True, 4).info_1d_plot(0, 'hor.', 0, "-", False, 0, 1).plot_1d(enable=False)
-#     img.image = eBeamTraj_srw["eTraj"]["By"]
-#     img.info_1d_plot(1, 'ver.', 1, "-", False, 0, 1).plot_1d(file_name=file_name, enable=True, hold=True)
-
-#     if integrals:
-        
-#         frst_field_integral_Bx = integrate.cumulative_trapezoid(eBeamTraj_srw["eTraj"]["Bx"],
-#                                                                 eBeamTraj_srw["eTraj"]["Z"], initial=0)
-#         scnd_field_integral_Bx = integrate.cumulative_trapezoid(frst_field_integral_Bx,
-#                                                                 eBeamTraj_srw["eTraj"]["Z"], initial=0)
-        
-#         frst_field_integral_By = integrate.cumulative_trapezoid(eBeamTraj_srw["eTraj"]["By"],
-#                                                                 eBeamTraj_srw["eTraj"]["Z"], initial=0)
-#         scnd_field_integral_By = integrate.cumulative_trapezoid(frst_field_integral_By,
-#                                                                 eBeamTraj_srw["eTraj"]["Z"], initial=0)
-        
-#         img = PlotManager(frst_field_integral_Bx,  eBeamTraj_srw["eTraj"]["Z"])
-#         img.additional_info('first field integral', "s [m]",  "[T$\cdot$m]", xmin=xmin, xmax=xmax, ymin=None, ymax=None)
-#         img.aesthetics(400, True, True, 0, 1, True, 4).info_1d_plot(2, 'hor.', 0, "-", False, 0, 1).plot_1d(enable=False)
-#         img.image = frst_field_integral_By
-#         img.info_1d_plot(3, 'ver.', 1, "-", False, 0, 1).plot_1d(file_name=file_name, enable=True, hold=True)
-
-#         img = PlotManager(scnd_field_integral_Bx,  eBeamTraj_srw["eTraj"]["Z"])
-#         img.additional_info('second field integral', "s [m]",  "[T$\cdot$m²]", xmin=xmin, xmax=xmax, ymin=None, ymax=None)
-#         img.aesthetics(400, True, True, 0, 1, True, 4).info_1d_plot(4, 'hor.', 0, "-", False, 0, 1).plot_1d(enable=False)
-#         img.image = scnd_field_integral_By
-#         img.info_1d_plot(5, 'ver.', 1, "-", False, 0, 1).plot_1d(file_name=file_name, enable=True, hold=True)
